Establish database connection

Establishes a connection with the WDFW Postgres database using local configuration and user credentials.

Usage

establish_db_con(max_attempts = 5, delay_seconds = 3)

Arguments

max_attempts

Numeric. Number of attempts to make when attempting to establish connection. Defaults to 5.

delay_seconds

Numeric. Number of seconds to wait between attempt. Defaults to 3.

Value

Connection to WDFW PostgreSQL database.

Details

requires a local config file that is on gitignore list Dev note: we could add template config file to the inst/ folder, and then add a function that sets up the config file.