plot paired census index

Pared down version of prep_inputs_pe_ang_hrs_vhcl_trlr.R function to return table with season-long sum of angler counts from paired census and index count surveys and scatter plot showing ratio of census to index counts relative to a 1:1 line in a scatterplot.

Usage

plot_paired_census_index(
  days,
  dwg_summarized,
  interview_ang_per_vehic,
  census_expan,
  angler_hours_daily_mean,
  params
)

Arguments

days

??

dwg_summarized

??

interview_ang_per_vehic

??

census_expan

??

angler_hours_daily_mean

??

params

??

Value

List. $plot is ggplot object, $census_TI_expan is ??

Details

Dev notes: where is angler_hours_daily_mean defined?