title

Calc fishing_time & total time conditional on use angler vs group counts

Usage

prep_dwg_interview_fishing_time(
  dwg_interview,
  min_fishing_time,
  study_design,
  person_count_type
)

Arguments

dwg_interview

interview data from dwg filtered using start & end dates passed from params

min_fishing_time

numeric passed from params to filter fishing_time at least as long as this per-person duration

study_design

string passed from params denoting which study design was followed during data collection

person_count_type

??

Value

??

Details

Dev note: relies on person_count_type in global.